Added identity mapping of all potential (64TB) RAM before multiboot parsing.
